Replace oh-my-zsh with starship and antibody
I've been frustrated with oh-my-zsh for a while now. The constant
updates (outside of any package manager), the heavy weight, the
obfuscation of everything going on in my actual shell configuration.
This commit removes it entirely, and replaces it with two smaller,
lighter-weight components.

Starship is used to replace the prompt configuration ("theme").
This is a small shell-agnostic rust binary which supports extensive
configuration options.

Antibody is used to replace the plugin loading. This is a small Go
binary which fetches plugins and produces a static bundles file to
source them all from. This keeps shell startup fast.

With each of these I've chosen a fairly minimal configuration, with
only a few of the Starship modules enabled, and only three plugins
loaded.

Hopefully I don't run into anything that oh-my-zsh was giving me
secretly and that I was relying on unconsciously, and that I have
no idea how to get back.